# salesforce-agentforce/examples/metadata/openapi_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Neo4j Knowledge Graph API",
    "description": "REST API for querying the Company News Knowledge Graph (250k entities). Exposes company profiles, news search, industry data, and graph analytics for Salesforce AgentForce External Service Actions.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://your-neo4j-bridge.example.com",
      "description": "Neo4j REST Bridge (update to your deployment URL)"
    }
  ],
  "paths": {
    "/research/company": {
      "post": {
        "operationId": "researchCompany",
        "summary": "Full company research in one call",
        "description": "Retrieves comprehensive company information including profile (industry, location, leadership), recent news headlines with sentiment scores, and organizational relationships. Use this as the primary research action when a user asks about a company. Returns company_id for follow-up queries.",
        "tags": ["Research"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["company_name"],
                "properties": {
                  "company_name": {
                    "type": "string",
                    "description": "Company name to research (e.g., 'Apple', 'Tesla')"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company research results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile": {
                      "type": "object",
                      "properties": {
                        "name": {"type": "string"},
                        "company_id": {"type": "string"},
                        "summary": {"type": "string"},
                        "locations": {"type": "array", "items": {"type": "string"}},
                        "industries": {"type": "array", "items": {"type": "string"}},
                        "leadership": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {"type": "string"},
                              "title": {"type": "string"}
                            }
                          }
                        }
                      }
                    },
                    "recent_news": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "article_id": {"type": "string"},
                          "title": {"type": "string"},
                          "date": {"type": "string"},
                          "sentiment": {"type": "number"}
                        }
                      }
                    },
                    "relationships": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "organization": {"type": "string"},
                          "relationship_type": {"type": "string"},
                          "distance": {"type": "integer"}
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Company not found"
          }
        }
      }
    },
    "/tools/search_companies": {
      "post": {
        "operationId": "searchCompanies",
        "summary": "Search for companies by name",
        "description": "Full-text search to find companies matching a name or keyword. Returns company names and IDs. Use when the exact company name is unknown.",
        "tags": ["Company"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["search"],
                "properties": {
                  "search": {
                    "type": "string",
                    "description": "Search term for company name"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of matching companies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "companies": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "company_id": {"type": "string"},
                          "name": {"type": "string"},
                          "summary": {"type": "string"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/list_industries": {
      "get": {
        "operationId": "listIndustries",
        "summary": "List all industry categories",
        "description": "Returns all available industry categories in the knowledge graph. Use before filtering companies by sector.",
        "tags": ["Industry"],
        "responses": {
          "200": {
            "description": "List of industry categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "industries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "industry": {"type": "string"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/companies_in_industry": {
      "post": {
        "operationId": "getCompaniesByIndustry",
        "summary": "Get companies in a specific industry",
        "description": "Returns companies in a given industry category. Use listIndustries first to get valid category names.",
        "tags": ["Industry"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["industry"],
                "properties": {
                  "industry": {
                    "type": "string",
                    "description": "Industry category name from listIndustries"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Companies in the industry",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "companies": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "company_id": {"type": "string"},
                          "name": {"type": "string"},
                          "summary": {"type": "string"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/find_influential_companies": {
      "post": {
        "operationId": "findInfluentialCompanies",
        "summary": "Find most influential companies by network analysis",
        "description": "Uses PageRank graph algorithm to identify most influential companies by organizational relationships. Higher score = more connected and important in the business network.",
        "tags": ["Analytics"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "integer",
                    "default": 10,
                    "description": "Number of top companies to return"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Top companies by influence score",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "companies": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "company_name": {"type": "string"},
                          "company_id": {"type": "string"},
                          "score": {"type": "number"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "healthCheck",
        "summary": "Check API health",
        "description": "Verify Neo4j connectivity and API status.",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "Health status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {"type": "string"},
                    "database": {"type": "string"},
                    "organization_count": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "API key configured in Salesforce Named Credential as custom header"
      }
    }
  },
  "security": [
    {"ApiKeyAuth": []}
  ]
}